Fix hash test results on big-endian systems. (#494)
The result is the same, but looking at the *raw* bytes, the results are
byte swapped. So comparing against a list of bytes from a little-endian
system won't work.

Just swap the byte order before comparing against the ground truth
values. This means there's no need to save two copies of the ground
truth.