Switch to stack-based protection in dictionary.c
`dict_init()` and `dict_free()` imply the dictionary can be
initialised and freed anywhere on the stack. `dict_free()` is removed
in favour of `PROTECT_DICT()` to make it clear it that protection and
unprotection must happen in stack consistent order. It takes a
protection counter, so a simple `UNPROTECT()` is sufficient.