Support Altvec objects in vec_slice
This dispatches to ALTVEC_EXTRACT_SUBSET for altvec objects,
falling back to the normal slicing code if the Altvec object does not
declare a extract_subset method.

For versions of R prior to the introduction of Altvec (< 3.5.0) stub
macros are defined resulting in a dead conditional, which should be
removed by dead code optimizer in the compiler.