Use dummy `verify_errors()` operator
For now it indicates that an error expectation should be updated in
sync with its output verification.

In the future, there could be a testthat implementation that pushes
automatically to the right txt file (inferred from the current
context, by default the file name) inside the right section (inferred
from the `test_that()` title).