Modularize TRUELENGTH saving/reallocating helpers
We only have to store the old TRUELENGTH value in `strings` / `truelengths` if R is actually using it, i.e. the value is >0, otherwise we can always reset it to the default value of 0. It is rare for R to use the TRUELENGTH, so we should rarely have to reallocate the `strings` and `truelengths` vectors. This saves us some time with large vectors when most of the strings are unique, because we previously saved every unique value and its TRUELENGTH.