Revise integer64 proxy (#1378)
* Implement alternative integer64 proxy based on the complex atomic type

* Replace complex proxy with a two column data frame proxy

Since this isn't used in `vec_proxy()`, we shouldn't suffer any performance hits for using a data frame proxy (i.e. `vec_chop()` still uses the double storage of the integer64, which is fast). Using a data frame proxy with double columns also makes the comparison proxy simpler, because double vectors can be compared, unlike complex vectors.

* NEWS

* complex -> data frame

* Use manually generated NA integer64 value

* Add a custom `as.data.frame.array()` that works with 1-D arrays

This is also different from `as.data.frame2()`, in that each column of the result has the opportunity to retain the input class.

* Test for NA-ness with `isnan()`

* Add a comment about UB in the cast

Since bit64 relies on this behavior, if it was ever working incorrectly then it would mean that bit64 is seriously broken.