Fix small buglet when `missing = "propagate"`
If doing both the lhs and rhs here, it was possible to overwrite `loc_upper_o_needles` when doing the lhs, then use that overwritten variable when doing the rhs, which is wrong. I'm pretty sure this can't happen in the current setup, but it is possible with some potential changes I am making to the computation of the joint ranks.