Always match missing values exactly
Previously we matched missing values exactly if `>=` or `<=` was used, but not if `>` or `<` was used. The rationale was that we were applying "or equal to" logic to match the missing values.

This proved to be a little annoying in practice, as we typically want missing values to match other missing values exactly, regardless of the comparison operation we are performing.

You could also argue that one missing value never "equals" another anyways, so trying to use the "or equal to" logic never really held correctly.