Speed up `multiple` for `"first"` and `"last"`
We have removed the requirement on "row order" when generating the nesting containers. This greatly improves the performance of the recursive function because we jump around less along `haystack`. This has the cost of having to go through most of the `"all"` code path to compactly record all matches, and then in `expand_compact_indices()` we filter down to the first/last match.

I've left stubs in to be able to support `multiple = "any"`, which should be even faster than first/last