Tweak how filter matches are recorded
This is a minor change that shouldn't really have much performance advantage (maybe a little). We previously recorded the upper match location in the actual haystack as the filter match, but the first match array always held the lower match location, so the `loc_first_match_haystack == loc_filter_match_haystack` condition in `expand_compact_indices()` was never getting run if there were duplicate matches. We now consistently record the lower match everywhere, and we also now record the lower match location in the haystack ordering vector rather than the location in the haystack itself, which is more consistent with how recording the first match location works and cleans up the code a little.