Implement `vec_locate_matches()` (#1472)
* Implement C side of `vec_locate_matches()`

* Implement R side of `vec_locate_matches()`

* Test `vec_locate_matches()`

* Update pkgdown reference

* NEWS bullet

* Relax type check so unspecified logicals don't look different

* Always match missing values exactly

Previously we matched missing values exactly if `>=` or `<=` was used, but not if `>` or `<` was used. The rationale was that we were applying "or equal to" logic to match the missing values.

This proved to be a little annoying in practice, as we typically want missing values to match other missing values exactly, regardless of the comparison operation we are performing.

You could also argue that one missing value never "equals" another anyways, so trying to use the "or equal to" logic never really held correctly.

* Reduce the mental overhead of `condition = NULL`

`expand_match_on_nothing()` went through a lot of extra effort to ensure it was correct with all combination of optional arguments (like `no_match` and `multiple`).

This commit allows us to remove this function in favor of letting vectors of `0` go through the main code path (so everything matches). It has negligable performance impact since there is only 1 unique value.

* `document()` the new missing value matching behavior

* Fix buglet when checking for multiple matches

* Speed up `multiple` for `"first"` and `"last"`

We have removed the requirement on "row order" when generating the nesting containers. This greatly improves the performance of the recursive function because we jump around less along `haystack`. This has the cost of having to go through most of the `"all"` code path to compactly record all matches, and then in `expand_compact_indices()` we filter down to the first/last match.

I've left stubs in to be able to support `multiple = "any"`, which should be even faster than first/last

* Update snapshot errors using dev rlang

* Implement `multiple = "any"`, which can be faster than `"first"/"last"`

Particularly useful for a `vec_detect_matches()` style wrapper, where you only need to know if there are any matches. This is what is done in `anti_join()` and `semi_join()`.

* Add `incomplete = "compare"` back in as the default

This was the previous default before I changed it to use `incomplete = "match"`. I think both are useful, but `"compare"` is probably more intuitive?

- `"compare"` is useful when you are checking if `x` precedes `y` with a `<` condition, in which case you probably don't want missing values to match.
- `"match"` is useful when you are checking if `[a, b)` overlaps `[c, d)` with `a < d, b > c`, in which case you probably do still want `[NA, NA)` to be considered a match.

* Update `r_poke_names()` to `r_attrib_poke_names()`

* Update `r_poke_class()` to `r_attrib_poke_class()`

* Use `r_lazy_null` in `vec_ptype()`

* Use `r_lazy_null` in `vec_cast()`

* Update snapshot tests

* Slightly revise `vec_joint_proxy_order()` documentation

* Remove ability to set `condition = NULL`

This is exactly equivalent to providing two integer vectors with the same size as `needles` and `haystack` that are filled with `1L`s. We will force dplyr to do this, and keep the `vec_locate_matches()` implementation simpler.

* Remove the NEWS bullet

* Remove `vec_locate_matches()` from the pkgdown reference

* Unexport `vec_locate_matches()`

* Add internal FAQ for `vec_locate_matches()` algorithm

* Provide an experimental shim of `vec_locate_matches()`

* Use modern `ffi_` prefix

* Use `r_arg_as_bool()`

* Tweak style

* Tweak how filter matches are recorded

This is a minor change that shouldn't really have much performance advantage (maybe a little). We previously recorded the upper match location in the actual haystack as the filter match, but the first match array always held the lower match location, so the `loc_first_match_haystack == loc_filter_match_haystack` condition in `expand_compact_indices()` was never getting run if there were duplicate matches. We now consistently record the lower match everywhere, and we also now record the lower match location in the haystack ordering vector rather than the location in the haystack itself, which is more consistent with how recording the first match location works and cleans up the code a little.

* Adjust documentation in `expand_compact_indices()`

* Use `int` for protection counter - as noticed by rchk

* Pass correct location through to `stop_matches_nothing()`

* Optimize first/last branch a bit

Brings the extreme case below down from 7.5s to 4.5s

```
x <- rep(1L, 1e5)
vec_locate_matches(x, x, multiple = "first")
```