Use alternate `any_multiple_needles` detection
We can't use the simple `loc < size_needles` check to see if we are in the "extra" match section. When a filter is involved, that isn't enough because we could have filtered out a match that occurred before the extra matches, so the extra match could actually be the first one.