Use a RAWSXP to store `strings` and `uniques` (#2091)
Both of these are subsets of the original `x` vector. We are not claiming ownership over them in any way, and `x` cannot outlive us, so we don't need to store them in a STRSXP or use `SET_STRING_ELT()` to track references (which we weren't even doing before, and it was clearly fine).